requirements.txt
bidict==0.22.0
click==8.1.1
dnspython==2.2.1
eventlet==0.33.0
Flask==2.1.1
Flask-JWT-Extended==4.3.1
Flask-SocketIO==5.1.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
protobuf==3.19.4
PyJWT==2.3.0
python-dotenv==0.20.0
python-engineio==4.3.1
python-socketio==5.5.2
six==1.16.0
Werkzeug==2.1.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
